Move update status logic in procedure plugin
