+++ nuxt/nuxt.config.js
@@ -51,6 +51,7 @@ export default {
     { src: '~/plugins/django-api.js', mode: 'client' }, // Need to be before urbanisator plugin
     { src: '~/plugins/collectivite.js', mode: 'client' },
     { src: '~/plugins/event.js', mode: 'client' },
+    { src: '~/plugins/procedure.js', mode: 'client' },
     { src: '~/plugins/urbanisator.js' },
     { src: '~/plugins/pdfMake.js', mode: 'client' },
     { src: '~/plugins/analytics.js' },
